requirements.txt
numpy
pandas
matplotlib
pillow
gradio
python-dotenv
google-genai
